- refactor files
- add viewModels delegate to remove boilerplate code
