+++ .idea/modules.xml
@@ -2,8 +2,8 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/Inventory-Tracker.iml" filepath="$PROJECT_DIR$/Inventory-Tracker.iml" />
-      <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/Inventory-Tracker.iml" filepath="$PROJECT_DIR$/.idea/Inventory-Tracker.iml" group="Inventory-Tracker" />
+      <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" group="Inventory-Tracker/app" />
     </modules>
   </component>
 </project>
