--- app/src/main/java/com/fazemeright/myinventorytracker/additem/AddItemViewModelFactory.kt
@@ -1,24 +1,0 @@
-package com.fazemeright.myinventorytracker.additem
-
-import android.app.Application
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.fazemeright.myinventorytracker.database.InventoryDatabase
-
-/**
- * This is pretty much boiler plate code for a ViewModel Factory.
- *
- * Provides the SleepDatabaseDao and context to the ViewModel.
- */
-class AddItemViewModelFactory(
-    private val dataSource: InventoryDatabase,
-    private val application: Application
-) : ViewModelProvider.Factory {
-    @Suppress("unchecked_cast")
-    override fun <T : ViewModel?> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(AddItemViewModel::class.java)) {
-            return AddItemViewModel(dataSource, application) as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class")
-    }
-}
