+++ app/src/main/java/com/fazemeright/myinventorytracker/database/inventoryitem/InventoryItem.kt
@@ -1,4 +1,4 @@
-package com.fazemeright.myinventorytracker.database
+package com.fazemeright.myinventorytracker.database.inventoryitem
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
