+++ app/src/main/java/com/fazemeright/myinventorytracker/ui/addbag/AddBagActivity.kt
@@ -1,38 +1,37 @@
-package com.fazemeright.myinventorytracker.addbag
+package com.fazemeright.myinventorytracker.ui.addbag
 
 import android.graphics.Color
 import android.os.Bundle
 import android.view.MenuItem
 import android.widget.Toast
+import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProviders
 import com.fazemeright.myinventorytracker.R
 import com.fazemeright.myinventorytracker.database.InventoryDatabase
 import com.fazemeright.myinventorytracker.databinding.ActivityAddBagBinding
+import com.fazemeright.myinventorytracker.ui.itemlist.BaseViewModelFactory
 import com.jaredrummler.android.colorpicker.ColorPickerDialog
 import com.jaredrummler.android.colorpicker.ColorPickerDialogListener
+
 
 class AddBagActivity : AppCompatActivity(), ColorPickerDialogListener {
 
     private var bagColor: Int = 0
     private val dialogId: Int = 1001
-    lateinit var viewModelFactory: AddBagViewModelFactory
-    lateinit var viewModel: AddBagViewModel
     lateinit var binding: ActivityAddBagBinding
+    val viewModel: AddBagViewModel by viewModels {
+        BaseViewModelFactory(
+            application = application,
+            dataSource = InventoryDatabase.getInstance(application)
+        )
+    }
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
         binding = DataBindingUtil.setContentView(this, R.layout.activity_add_bag)
-
-        val application = requireNotNull(this).application
-
-        val dataSource = InventoryDatabase.getInstance(application)
-
-        viewModelFactory = AddBagViewModelFactory(dataSource, application)
-
-        viewModel = ViewModelProviders.of(this, viewModelFactory).get(AddBagViewModel::class.java)
 
         binding.viewModel = viewModel
 
@@ -54,7 +53,11 @@ class AddBagActivity : AppCompatActivity(), ColorPickerDialogListener {
         }
 
         binding.btnAddBag.setOnClickListener {
-            viewModel.onAddBagClicked(binding.edtBagName.text.toString(), bagColor, binding.edtBagDesc.text.toString())
+            viewModel.onAddBagClicked(
+                binding.edtBagName.text.toString(),
+                bagColor,
+                binding.edtBagDesc.text.toString()
+            )
         }
 
         viewModel.navigateBackToItemList.observe(this, Observer { navigate ->
@@ -71,7 +74,8 @@ class AddBagActivity : AppCompatActivity(), ColorPickerDialogListener {
     }
 
     override fun onColorSelected(dialogId: Int, color: Int) {
-        Toast.makeText(this, "Selected Color: #" + Integer.toHexString(color), Toast.LENGTH_SHORT).show()
+        Toast.makeText(this, "Selected Color: #" + Integer.toHexString(color), Toast.LENGTH_SHORT)
+            .show()
         binding.viewBagColorDisplay.setBackgroundColor(color)
         bagColor = color
 //        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
