+++ app/src/main/java/com/fazemeright/myinventorytracker/ui/itemdetail/ItemDetailViewModelFactory.kt
@@ -1,23 +1,20 @@
-package com.fazemeright.myinventorytracker.addbag
+package com.fazemeright.myinventorytracker.ui.itemdetail
 
 import android.app.Application
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.fazemeright.myinventorytracker.database.InventoryDatabase
+import com.fazemeright.myinventorytracker.database.inventoryitem.InventoryItemDao
 
-/**
- * This is pretty much boiler plate code for a ViewModel Factory.
- *
- * Provides the SleepDatabaseDao and context to the ViewModel.
- */
-class AddBagViewModelFactory(
+class ItemDetailViewModelFactory(
     private val dataSource: InventoryDatabase,
-    private val application: Application
+    private val application: Application,
+    private val itemInBag: InventoryItemDao.ItemInBag
 ) : ViewModelProvider.Factory {
     @Suppress("unchecked_cast")
     override fun <T : ViewModel?> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(AddBagViewModel::class.java)) {
-            return AddBagViewModel(dataSource, application) as T
+        if (modelClass.isAssignableFrom(ItemDetailViewModel::class.java)) {
+            return ItemDetailViewModel(dataSource, application, itemInBag) as T
         }
         throw IllegalArgumentException("Unknown ViewModel class")
     }
