+++ app/src/main/res/layout/activity_add_item.xml
@@ -7,7 +7,7 @@
 
         <variable
                 name="viewModel"
-                type="com.fazemeright.myinventorytracker.additem.AddItemViewModel" />
+                type="com.fazemeright.myinventorytracker.ui.additem.AddItemViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
