+++ app/src/main/res/layout/activity_item_detail.xml
@@ -7,11 +7,11 @@
 
         <variable
                 name="viewModel"
-                type="com.fazemeright.myinventorytracker.itemdetail.ItemDetailViewModel" />
+                type="com.fazemeright.myinventorytracker.ui.itemdetail.ItemDetailViewModel" />
 
         <variable
                 name="item"
-                type="com.fazemeright.myinventorytracker.database.InventoryItemDao.ItemInBag" />
+                type="com.fazemeright.myinventorytracker.database.inventoryitem.InventoryItemDao.ItemInBag" />
 
     </data>
